+++ controllers/user/error.go
@@ -3,15 +3,8 @@ package user
 import "errors"
 
 var (
-	// ErrUserDoesNotExist mean the user already exist in DB
-	ErrUserDoesNotExist = errors.New("user does not exist")
-
-	// ErrUserAlreadyExist mean the user already exist in DB
-	ErrUserAlreadyExist = errors.New("user already exist")
-
-	// ErrMissingInformation mean the information user given missing
-	ErrMissingInformation = errors.New("missing information. please check again")
-
-	// ErrPasswordContainWhiteSpace mean the password should not contain white space
+	ErrUserDoesNotExist          = errors.New("user does not exist")
+	ErrUserAlreadyExist          = errors.New("user already exist")
+	ErrMissingInformation        = errors.New("missing information. please check again")
 	ErrPasswordContainWhiteSpace = errors.New("password should not contain white space")
 )
