+++ handlers/new.go
@@ -4,12 +4,6 @@ import (
 	"github.com/phuc-create/go-simple-crud/controllers/user"
 )
 
-type UserInput struct {
-	ID       string `json:"id"`
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
-
 type Handler struct {
 	userServices user.Controllers
 }
