+++ handlers/user.go
@@ -3,12 +3,25 @@ package handlers
 import (
 	"encoding/json"
 	"errors"
-	"github.com/go-chi/chi/v5"
 	"github.com/phuc-create/go-simple-crud/helpers"
 	"github.com/phuc-create/go-simple-crud/models"
 	"net/http"
 	"time"
 )
+
+type UserInput struct {
+	ID       string `json:"id"`
+	Username string `json:"username"`
+	Password string `json:"password"`
+}
+
+type UserResponse struct {
+	ID        string    `json:"id"`
+	Username  string    `json:"username"`
+	Password  string    `json:"password"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+}
 
 func validateInfoUser(user UserInput) error {
 	if user.ID == "" {
@@ -26,26 +39,26 @@ func validateInfoUser(user UserInput) error {
 }
 
 // GetUserByID return user following id
-func (h Handler) GetUserByID(w http.ResponseWriter, r *http.Request) {
-	userID := chi.URLParam(r, "userID")
-	if userID == "" {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, "User Not Found!Pls check again.")
-		return
-	}
-	user, err := h.userServices.GetUserByID(userID)
-	if err != nil {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	helpers.ResponseWithJSON(w, http.StatusOK, user)
-}
+//func (h Handler) GetUserByID(w http.ResponseWriter, r *http.Request) {
+//	userID := chi.URLParam(r, "userID")
+//	if userID == "" {
+//		helpers.ResponseWithErrs(w, http.StatusBadRequest, "Invalid user ID!Pls check again.")
+//		return
+//	}
+//	user, err := h.userServices.GetUserByID(userID)
+//	if err != nil {
+//		helpers.ResponseWithErrs(w, http.StatusInternalServerError, err.Error())
+//		return
+//	}
+//	helpers.ResponseWithJSON(w, http.StatusOK, user)
+//}
 
 // GetAllUser return all users available in DB
 func (h Handler) GetAllUser(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	list, _ := h.userServices.GetAllUser()
 	if len(list) < 1 {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, "Could not find any user!")
+		helpers.ResponseWithErrs(w, http.StatusInternalServerError, "Could not find any user!")
 	}
 	helpers.ResponseWithJSON(w, http.StatusOK, list)
 }
@@ -66,53 +79,53 @@ func (h Handler) CreateUser(w http.ResponseWriter, r *http.Request) {
 		UpdatedAt: time.Now(),
 	})
 	if err != nil {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
+		helpers.ResponseWithErrs(w, http.StatusInternalServerError, err.Error())
 		return
 	}
 	helpers.ResponseWithJSON(w, http.StatusOK, newUser)
 }
 
 // DeleteUser delete a user
-func (h Handler) DeleteUser(w http.ResponseWriter, r *http.Request) {
-	userID := chi.URLParam(r, "userID")
-	if userID == "" {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, "User Not Found!Pls check again.")
-		return
-	}
-	result, err := h.userServices.DeleteUser(userID)
-
-	if err != nil {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	helpers.ResponseWithJSON(w, http.StatusOK, result)
-}
+//func (h Handler) DeleteUser(w http.ResponseWriter, r *http.Request) {
+//	userID := chi.URLParam(r, "userID")
+//	if userID == "" {
+//		helpers.ResponseWithErrs(w, http.StatusBadRequest, "User Not Found!Pls check again.")
+//		return
+//	}
+//	result, err := h.userServices.DeleteUser(userID)
+//
+//	if err != nil {
+//		helpers.ResponseWithErrs(w, http.StatusInternalServerError, err.Error())
+//		return
+//	}
+//	helpers.ResponseWithJSON(w, http.StatusOK, result)
+//}
 
 // UpdateUserByID update user following ID
-func (h Handler) UpdateUserByID(w http.ResponseWriter, r *http.Request) {
-	var user UserInput
-	userID := chi.URLParam(r, "userID")
-	if err := json.NewDecoder(r.Body).Decode(&user); err != nil {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	if err := validateInfoUser(UserInput{
-		ID:       userID,
-		Username: user.Username,
-		Password: user.Password,
-	}); err != nil {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	result, err := h.userServices.UpdateUserByID(&models.User{
-		ID:       userID,
-		Username: user.Username,
-		Password: user.Password,
-	})
-
-	if err != nil {
-		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	helpers.ResponseWithJSON(w, http.StatusOK, result)
-}
+//func (h Handler) UpdateUserByID(w http.ResponseWriter, r *http.Request) {
+//	var user UserInput
+//	userID := chi.URLParam(r, "userID")
+//	if err := json.NewDecoder(r.Body).Decode(&user); err != nil {
+//		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
+//		return
+//	}
+//	if err := validateInfoUser(UserInput{
+//		ID:       userID,
+//		Username: user.Username,
+//		Password: user.Password,
+//	}); err != nil {
+//		helpers.ResponseWithErrs(w, http.StatusBadRequest, err.Error())
+//		return
+//	}
+//	result, err := h.userServices.UpdateUserByID(&models.User{
+//		ID:       userID,
+//		Username: user.Username,
+//		Password: user.Password,
+//	})
+//
+//	if err != nil {
+//		helpers.ResponseWithErrs(w, http.StatusInternalServerError, err.Error())
+//		return
+//	}
+//	helpers.ResponseWithJSON(w, http.StatusOK, result)
+//}
